pyproject.toml
# https://python-poetry.org/docs/libraries/#versioning
# https://pypi.org/project/hcaptcha-challenger/#history

[tool.poetry]
# python -m build && twine upload dist/*
name = "dify-knowledge-pipeline"
version = "0.1.7"
description = "🥂 Gracefully embedding multimodal-knowledge to Dify"
license = "Apache License 2.0"
authors = ["QIN2DIM <yaoqinse@gmail.com>", "Bingjie Yan <bj.yan.pa@qq.com>"]
readme = "README.md"
homepage = "https://github.com/QIN2DIM/dify-knowledge-pipeline"
repository = "https://github.com/QIN2DIM/dify-knowledge-pipeline"
documentation = "https://github.com/QIN2DIM/dify-knowledge-pipeline"
keywords = ["dify", "RAG", "dify-knowledge-pipeline", "multimodal-knowledge"]
packages = [{ include = "dify_knowledge_pipeline", format = "sdist" }]
classifiers = [
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
]

[virtualenvs]
# https://python-poetry.org/docs/configuration/#virtualenvsin-project
# Create a `.venv` virtual environment in the project root
in-project = true

# https://python-poetry.org/docs/pyproject/#dependencies-and-dependency-groups
[tool.poetry.dependencies]
python = "^3.10"

# Run `poetry install` in the project root
loguru = "*"
pydantic = "^2"
tiktoken = "^0.7.0"
tqdm = "*"
python-dotenv = "*"
langchain_text_splitters = "^0.2"
httpx = "*"

[tool.poetry.group.dev.dependencies]
# https://docs.pytest.org/en/stable/reference/plugin_list.html#plugin-list
# https://docs.pytest.org/en/stable/contents.html
# https://pytest-asyncio.readthedocs.io/en/latest/
pytest = "*"
pytest-asyncio = "*"
# https://black.readthedocs.io/en/stable/
black = "^24.4.2"

twine = "^5.1.0"
[tool.pytest.ini_options]
# https://docs.pytest.org/en/stable/reference/reference.html#configuration-options
testpaths = ["tests", "examples"]
asyncio_mode = "auto"
filterwarnings = "ignore::DeprecationWarning"

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
skip-magic-trailing-comma = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
